rev 3A (lv): preliminary circuit
it's basically the original low-voltage circuit, but with another
mosfet to control the second EXT pin, with a schottky in parallel
to reduce Vf across the mosfet.

it should work, but it's sad that it has to be low voltage. it also
requires an additional extended part for a smaller footprint mosfet
because no way can we fit 3 SOT-23s on the board currently.